update config for multi tenancy, add controller for login auth and logout

+++ src/main/java/com/yamidev/multitenancy/dto/entity/UserLogin.java
@@ -13,5 +13,5 @@ public class UserLogin implements Serializable {
 
     private String userName;
     private String password;
-    private Integer tenantOrClientId;
+    private Long tenantOrClientId;
 }
